[Resource] ApplicationsCount() refactoring
The method has been refactored moving the most of the code into a
private overloading method. This one returns the size of the map of
App/EXC and a reference to map itself. The latter information has been
thought for an internal usage purpose. This is why this refactoring has
been performed.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>