[Config] Update default logleves configuration
The NOTICE loglevel is used as a "less-verbose" info level, e.g. for demo
purposes. This patch switch to this level all the BBQ modules which are
used to dump statistics on console, while keeping all other modules silent
at ERROR level.

This should be a quite clear configuration for demo purposes using the
RELEASE build version.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>