[Metrics] Added new PERIOD metric class
This new metrics class provides the basic support to collect and compute
statistics on time periods. An high-resolution timer is managed internally
by the class so that by simply calling the PeridoSample method we are
able to collect the time since previous time this method was called and
thus compute (and return) the current period.
Such metrics could be used to compute the frequency of an event.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>