[WorkingMode] ResourceUsageValue() refactoring
The ResourceUsage to retrieve can be referenced by using either a resource
"template path" or an "ID-based path". In the first case the value is taken
from the object referenced into the 'rsrc_usages' map (the map of built
from the recipe).
The latter instead check if a binding has been performed. In the positive
case it gets the value from the object referenced into the binding map
('sys_rsrc_usages'). Otherwise the recipe-based map would be considered.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>