[RA] SetView() refactoring
The default system state view now is now controlled in a centralized way by
the ResourceAccounter. More precisely, this behavior will be effective only
after the "Resource refactoring" patch.
The view is no more set for each resource used. The Resource itself will
ask to the RA which is the current system view before access a state view.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>