[Resource] Overall refactoring
Most of code has been moved from resource.h to the new resource.cc file.
This has been done to improve the maintainability and readability of the
code, along with its readability.

In this process, some useless method have been deprecated and removed.

Moreover, regarding the "default state view" a new behavior has been
implemented. Indeed this is retrieved by query the ResourceAccounter. It
keeps track of a "centralized" token, which references the default/system
state view for all the resources. This change reduces the effort in
preserving consistent views among the resources.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>