[AP] Bugfixed missing locking
Once a new message has been received from applications, the corresponding
command session (pcs) must be notified to wakeup the waiting thread.
A proper notification of a sleeping thread must grant that the thread to be
notified has already entered the waiting state, and this could be grant by
using a lock on a command session specific mutex.

Unreasonably, this mutex was in place as well as the locking, but this last
has been commented out... thus resulting on missing notification especially
on single/dual-core machines.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>