[Testing] Add profiling script
This patch provides a testing script which could be used to run a BBQ profilig
session. The profiling is performed by spawning:
- running a Barbeuqe RTRM instance with a properly configured set of TPD
- spawning a sequence of BbqRTLibTestApp with a configurable set of EXC

All the profiling setting are configurable by means of shell variables to be
defined on the script starting command line. For example:
CLUSTERS=4 PES="4 8" EXCS="10 20 30" ./profile.sh
will run a set of simulation which correspond to all the possible combination
if this three params, i.e running a 10 (20 and 30) EXCs test-app on a test
platform with 4 cluster counting 4 (8) PEs each one.

At the end of each TestApp execution, BBQ collected metrics are dumped and
organized into ".dat" output datafile, which are ready to be graphed.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>