[Recipe] NormalizeValues() refactoring
The function has been refactored to improve readability and efficiency.

Moreover the case of recipe with a single AWM is now managed. Indeed in
this case, maximum and minimum value coincide (delta = 0). Although we
don't have to perform a "penalization", but directly set the value of
the only AWM to 1.

Conversely, if delta = 0 and the AWMs are more than one, the recipe is
penalized by setting all the AWMs values to 0.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>