[RTLIB] Add Notification and Monitoring interface
This patch add a set of RTLib callbacks function which are intended to be
used by an application to notify the library about its run-time activities.
Right now we identified the activities defined by the Abstract Execution Model
defined by the BbqueEXC base class, i.e. Configuration, Processing (aka Run),
Monitoring and Suspension.

This hooks are expected to be implemented by the RTLib core to collect activity and
overheads measurements, for example exploiting target system performance
counters.

As a consequence of this patch, the RTLib version has been bumped up to
be v1.1

Signed-off-by: Patrick Bellasi <derkling@gmail.com>