[Timer] Switch to CLOCK_MONOTONIC for interval measurements
CLOCK_REALTIME represents the machine's best-guess as to the current
wall-clock, time-of-day time. As Ignacio and MarkR say, this means that
CLOCK_REALTIME can jump forwards and backwards as the system time-of-day
clock is changed, including by NTP.

CLOCK_MONOTONIC represents the absolute elapsed wall-clock time since some
arbitrary, fixed point in the past. It isn't affected by changes in the
system time-of-day clock.

To compute elapsed time between two events observed on the one machine without
an intervening reboot, CLOCK_MONOTONIC is the best option.


Signed-off-by: Patrick Bellasi <derkling@gmail.com>