[Application] AWM constraints refactoring
This patch acts on AWM constraints assertion methods, i.e.
SetWorkingModeConstraint() and ClearWorkingModeConstraints().

Set a lower bound means that all the AWMs having an ID lesser than the
one specified in the method call will be disabled. Thus they will not
considered by the scheduler. Similarly, but accordingly to the reverse
logic, if an upper bound is set: all the working modes with ID greater
to the one specified will be disabled.  An "exact value" constraint acts
in addictive way. This means that the working mode specified is added to
the constraint range (if not included yet).

Whenever a constraint of the same type occurs, the previous constraint
is replaced. This means that if the application sets a lower bound, a
second lower bound assertion will "overwrite" the firsts. The same for
upper bound case.

All the enabled AWM are now stored into a vector indexed by the ID of
the AWMs, in order to better handle the ranges of AWMs.

The methods have been purged from the ApplicationConfIF, since we do not
expect to make this API available to any plugins.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>