[Resource] Overloading of method Availability()
The overloaded method allows to know the amount of available resource
without consider the amount of resource used by the application specified.
This is useful whenever we are trying to allocate resources to an
Application/EXC which is already using some. The idea is that this should
contribute to prevent "avoidable migrations" due to the scheduling
decisions.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>