[RTLIB] Fix some notifications
The PostRun notification should not be called the last time we return from
BbqueEXC::onRun(). Indeed, the AEM API requires the application to issue
an "empty" call to this method to signal the completion of the workload
processing and thus allowing a proper collection of statistics on each previuos
processing call.

The Init() notification has been placed in between the completion of the
used code initialization (i.e. after the BbqueEXC::onSetup() call) and immediately
before to enter the control loop.

Finally, the Exit() notification has been placed immediately before the control
loop termination.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>