[Perf] Add support to access performance counters
This patch provides a first simple class to abstract access to Linux performance
counters. The Perf class exposes a set of methods to configure and manage
performance counters.

A client class could use this new utility to "enqueue" performance counters
in kernel space, enable/disable them and read their values. All the counters
added to an instance of this class are enabled/disabled together.

The implementation has been inspired from the "perf" tool provided with the
Linux kernel sources. Special credits for this goes to:
  Thomas Gleixner and Ingo Molnar
and all the other guys which contributed to the Linux Performance events
framework.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>