[RTLIB] Add support for Run-Time configurability
This patch add the possibility to configure and tune some RTLib behaviors
based on the configuration of the execution environment.
Specifically, a set of configuration options has already been added to tune
the behvaior of the RTLib concerning the usage of Performance Counters.

An environment variable "BBQUE_RTLIB_OPTS" could be defined to specify the
performance counter configuration to use. Such a configuration is specified
with a string of column separated single-char switches.
The switches supported so fare are:
- b: enabling usage of "big numbers" notations for the produced output
- c: enabling CSV output
- G: enabling global statistics collection
- K: disable Kernel and Hipervisor from collected statistics
- O: collect statistics on RTLIB overheads
- P[n]: enabling perf, with counters collection detailed specified by [n]
- s: set the CSV separator to use in CSV output mode

Signed-off-by: Patrick Bellasi <derkling@gmail.com>