[RTLIB] Add (pre) Setup notification
This patch introduce a new onSetup() method to the BbqueEXC base class.
Such a method is intended to host all the initialization code of a BBQ
application and serve to two main purposes:
1. it provides a cleaner interface to the application developer, which is
based just on call-back methods. Indeed, the developer is now encouraged to
put all application specific code into the provided based class methods, thus
avoiding the usage of the class constructor.
2. it provides an hook for the RTLib to properly initialize services before the
user-code is initialized, e.g. setup performances counters before user code
create some new threads.

The plain RTLib API introduces a new Setup notifier callback which is intended
to be called in early initialization stages, especially before the user application
initialization code. For example, this is a suitable place-holder for performance
monitoring code initialization which is expected to be merged for a following
patch.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>