[RTLIB] Bugfix EXC suspend/resume
This patch adds a proper support to SUSPEND a running EXC, by moving it into
a "wait for resources assignment" state and than resuming it when a new
valid AWM has been selected for it.

TODO: this management of the RTLib state is based on a partial state machine
which is quite hard to maintain due to the number of state flags which are
used. We should switch to a more formally defined FSM, with single state variable,
which should provide also a better dupport for testing/validation, run-time
verification and also being self-documenting.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>