[SchedPol] Return token view for scheduled resources
A scheduling policy should try to allocate resources to applications by
building a new view on the available resources. On a successful scheduling,
this view token should be reported to the caller process.

This patch set up the ground to support the SchedulerManager during the platform
synchronization step, where each application should be actually assigned the
set of scheduled resources.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>