[RA] Keep track of last scheduled view
Schedule plugins are in charge to identify a new possible schedule of available
resources on ready applications. Such a schedule is defined into a new resources
view and could be required by other Barbeque core modules, e.g. the SchedulerManager.

This patch provides the required support to keep track of the last successful
resources scheduling, as well as the accessories methods to set and get this info.
Moreover, once a SyncCommit is issued, thus the scheduled view has becomed the
current system view, the last scheduled view is properly released.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>