[RM] Link PlatformProxy module
The ResourceManager exploits now the PlatformProxy, which is now in charge to
load the platform data, i.e. to scan for available resources on the specific
target platform.

The TestPlatformData are also removed from the Barbeque initialization code.
Indeed, the PlatformProxy already provide support to load test platform data
if required (by default in the current implementation) instead of loading
the target platform real resources.

Finally, the PlatormProxy monitoring thread is also started. This thread
will be in charge to monitor resources state and to trigger proper actions
within the BBQ.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>