[Resources] Removed Available() overloading
An overloading of method Available() in Resource has been removed.
The version of the method was the one having a single argument
(AppPtr_t). The only version remained has been refactored by setting
the default value of the AppPtr_t argument to "null". This header
supports both the invocation where an AppPtr_t is specified than the
one without. Therefore the previous overloading became useless.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>