[Resources] class ResourceUsage new definition
"ResourceUsage" has been redefined as class, with several new
methods added:

- GetAmount() replaces direct reference to attribute 'value'
- GetBindingList() replaces direct reference to attribute 'binds'
- GetFirstResource() and GetNextResource() support iteration
among all the resource binding to which the required resource is
linked. Using this couple of methods ensures that only the
resource binding used by the given application are referenced.
- TrackFirstResource() and TrackLastResource() should be
called in the resource booking process, to keep track of the
resource bindings defined by the scheduling.

Note: direct references to the class attributes (from other
classes) are deprecated. Therefore suitable patches will be
introduced to exploit the new API.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>