[RA] Resource booking refactoring
The resource booking now behaves according to two different
approaches, distinguishing between the scheduling and the
synchronization step.

If a booking is due to the scheduling process, the resources are
allocated as usual: The amount of required resource is acquired
by checking the availability from the list of all the possible
bindings in the ResourceUsage object. Thus the amount is splitted
among the resource bindings (whenever a single resource binding
cannot satisfy the whole request).

If the booking is called during a synchronization instead, the RA
check the amount of resource allocated by the scheduling policy
("sch_view_token") and performs the effective acquisition.

This mechanism has been introduced to guarantee consistency
between the resource partitioning defined by the scheduling policy
and the real allocation committed during the synchronization.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>