[WorkingMode] New resource binding interface
A new data member has been added to the class: UsagesMapPtr_t sched_binds.

"sched_binds" points a map of ResourceUsage object which is built through
BindResource() calls. In the previous version of this method, such pointer
was provided as an argument. In the new one instead it is stored in
"sched_binds".

GetSchedResourceBinding() returns this pointer to map. The idea is that
this is the map that the scheduling policy builds to bind the resources
required in the recipe to the resources available in the system.

SetSchedResourceBinding() replaces the old SetResourceBinding(). It ends
the resource binding process by setting "sched_binds" as the set of
resources to allocate during the synchronization step.

GetResourceBinding() returns a pointer to the map of resource usages
to allocate to the application, according to the given working mode. It
is commonly called by the ResourceAccounter, during a synchronization
session. This method has not been changed.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>