[WorkingMode] New API for data member 'value'
The AWM 'value' has turned into a WorkingModeValue data type structure.
The struct stores the value parsed from the recipe and the normalized value
to use in the scheduling policy.

The write access to this data is provided by methods SetRecipeValue() and
SetNormalValue() respectively. The read access instead is given by
RecipeValue() and Value().

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>