[Recipe] Introduced control over the AWM IDs
The AWM IDs must be sequentially numbered from 0 to N. A recipe violating
this constraint should be rejected. This has been introduced to better
support constraint assertions on the AWMs set.

Furthermore, AWM descriptors are now stored into a vector instead of a map.
This provide more efficiency in inserting and searching.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>