[RTLIB_FIFO] Add macro to send custom sized messages
This patch introduces a more generic RPC_FIFO_SEND macro which allows
to send messages of specified size. This support could be used by messages
which allows a variable number of element to be send. For example the
SetConstraints call referes to a "basic" message defined by a struct but than
allows to push more constraints following this initial struct by simply
enqueuing them. Thus the need to specify explicitly the total size of the message
being sent over the FIFO channel.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>