[RM] Add Forced Reschedule EVENT
An event has been added to ask the ResourceManager for a forced re-scheduling.
This event could be triggered by other BarbequeRTRM code modules, for instance
after a re-definition of AWMs constraints for an application, in order to
schedule a new optimization.

TODO: add a support to schedule an Optimization after a specified Timeout as
described by issue ticket #13.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>