[RTLIB_FIFO] Bugfix to send multi-constraint messages
Multiple constraints could be asserted concurrently, this requires to
send a FIFO messages which is composed by a "basic" rpc_fifo_EXC_SET_t,
which has room for just a single constainta, followed by all the other N-1
constraints within the array.

Thus, this requires to send a buffer with variable and call-time defined size.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>