[AWM] Introduce struct 'resources'
A data structure (of type ResourceUsagesInfo) has been introduced:
"resources". This embodies the data members of type UsageMap_t or
UsageMapPtr_t in order to improve the readability and the location of the
information about resources in the class. Therefore data have been
re-arranged in the following "resources" attributes:

- from_recp: [was "recp_usages") Stores the map of resource requirements
parsed from the recipe
- on_sched: [was "sched_binds"] Is the map built by the scheduling policy
through the binding process
- to_sync: [was "sys_usages"] The map of resource usages ready to be
synchronized. We it is set, the scheduling request has been performed.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>