[AWM] Multiple resource binding support
A scheduling policy can now perform more than one resource binding per
time. This can be handled by specifying the "binding id" (bid) parameter in
BindResource(). The parameter is set to 0 as default, for the sake of
flexibility and retro-compatibility, whatever a policy does not aim to
exploit this feature.

Under the hood, what happens is that several binding maps are built. To
retrieve the specific binding all we need is to specify the "bid"
paramenter when calling GetSchedResourceBinding().

At the end, the resource binding to select for the synchronization step
MUST be one. Therefore the "bid" parameter has been added to
SetResourceBinding() too.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>