[Utility] Add generic formatting marker
This patch provides a generic formatting marker to log INFO sentences.
Such a marker is intended to be used by all BBQ code modules when a suitable
logger module could not be used (e.g. during early initialization).

Some cosmetics have been applied also to the BBQUE_FMT macro, which now
colors only the message and not the initial time and module token.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>