[PP_LINUX] Add memory nodes assignement
Right now we don't have control on the amount of memory a process can use.
However, we can force the memory nodes an application can use, which on a
generic Linux SMP machine, are represented by the socket ID where a process
has been scheduled. Indeed, here the socket ID is also associated to a single
memory node.

This patch removes the not yet supported memory quantity assignment and
replace it with a memory node constraint, which is managed by the same
"cpuset" controller used for CPUs assignment.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>