[AWM] Bugfix in clusters map construction
When the SetResourceBinding() is called, the map of clusters used in the
AWM is updated. A bug previously introduced does not let to perform this
update correctly, since it tried to get the cluster's ID from a "resource
template path" (where no ID are specified). This patch fixes such
misbehavior.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>