[RTLIB_Monitors] Loops optimization
According to me, the use of "result" in these snippets is quite sub-optimal.

In the first one, you loop on a TRUE condition and ther is not need to keep
track of the previous state. Thus:
   result = result && ...
is useless.

In the second, you use a for instead of a while, thus producing "code entropy"
when it is not needed at all. This could be coded with a while as in the previous
case. Moreover, blame me if I'm wrong, but again you are looking for
a FALSE condition (because of the &&) thus you could avoid the complete
goals checking by exiting as far as you get the first FALSE result.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>