[RTLIB_Monitors] CRITICAL, initialization by copy overhead
It seem to me that, as described by the comment provided by this patch,
that you improperly used construction-by-copy of targets vectors.
This means that a locally allocated vector should be copied multiple times
with a corresponding overheads that would be nice to profile ;-)

The suggestion is to make a proper use of a shared_pointer, allocate these
vectors on the heap and then pass only reference-countere pointers to the
inner classes.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>