[OperatingPoints] Introduced a new functionality in bbque_monitors
This commits introduces a new functionality inside the monitors library: a
manager of operating points.
This manager is used to create a structure with all the operating points
contained in an xml file and order it according to different metrics.
Each operating point is defined by its own application parameters and a set
of metrics that describe the performance/behavior obtained using that set.
Moreover, it is possible to define a sorting order for metrics.
Using a list of these elements, it's possible to define a full ordering of
operating points, useful for runtime management purposes.
When such a structure is created, it is possible to move between different operating
points according to priorities and filters given by the user.
A general filter on operating points is also provided within this commit.
Each filter consist of the name of a metric, a comparison function and a
value (eg. "Power" LessOrEqual 10). Using a list of filters such as
OP_FilterList, it's possible to filter out from a set of operating poins the
unwanted ones.

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Reviewed-by: Andrea Di Gesare <andrea.digesare@gmail.com>