[RTLIB] Add Pre/PostResume notifiers
This patch add a new pair of RTLib callbacks function which are intended to be
used by an application to notify the library about its suspend-resume activities.

This hooks are expected to be implemented by the RTLib core to collect activity
and overheads measurements, for example exploiting target system performance
counters.

    As a consequence of this patch, the RTLib version has been bumped up to
    be v1.2

Signed-off-by: Patrick Bellasi <derkling@gmail.com>