[GoalGap 1/8] [RPC] Add new RPC message
This patch add a new RPC message to allows an Execution Context (EXC) to
notify the RTRM about a "Goal Gap" (GG).
An EXC could assert a GG relative to the current assigned AWM if the
expected QoS level is lower that its actual value. Thus, the Barbeque RTRM
could do the best to schedule for this EXC a working mode which tries to
assign more resources to to its in order to compensate the asserted gap.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>