[GoalGap 6/8] [AM] Set Goal-Gap on the application
The specified Goal-Gap is configure for the specified Application.

TODO: Once a goal-gap is asserted a rescheduling could be triggered.
However, it is worth to have this rescheduling to be force only if the gap
is crossing a predefined threshold value or based on a suitable policy
decision. Right now we trigger it on a fixed based (20% gap). This should
be improved in future versions.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>