[RTLIB_Monitors] Bug fix in Generic Window
This commit fixes a bug that shows up when Generic Windows with unsigned values
are used. If a gap is going to be negative then the result is wrong because of the
operands are unsigned values. This commit forces a promotion to double for the
result of the goal evaluation and the goal value. This casted values are used just to
calculate the gaps while the original values are used to verify if the goal has been
achieved or not.

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Reviewed-by: Andrea Di Gesare <andrea.digesare@gmail.com>