[Application] UsageOutOfBounds e rsrc_constraints class members
The method checks if a given AWM violates some resource constraint.
Previously it was implemented as a static function, in order to exploit
the efficient "remove_if" method from STL class "list". However, since the
function requires access to the map of resource constraints
(rsrc_constraints), this data structure could be included as a private
member of class Application. This patch fixes this issue by abandoning the
usage of list::remove_if and moving back UsageOutOfBounds and
rsrc_constraints as class members.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>