[Resources] Change type of Application reference
The reference to the Applications/EXCs using the resource has been changed
both in classes Resource and ResourceUsage. Reference of type AppPtr_t
have been replaced by AppSPtr_t, which point to the read-only superclass
of Application. This has been done in order to restrict the access to
"write" methods that such classes do not need. This also provides a sort of
protection mechanism, improving security features.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>