[PP_LINUX] Add parsing of MEMORY resources
This patch add support to exploit the memory controller, which is now used to
parse MEMORY resources assigned to each NODE control group.
The amount of memory assigned to each attribute:
bbque/res/nodeX/memory.limit_in_bytes
is used to register that amount of memory into the ResourceAccounter as a
local cluster memory.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>