[RA] Number of resources and resource types support
A couple of methods have been introduced to return the number of resources
and resource types (or class) managed (registered). Information are stored
into a map, using a string object as key (the resource type) and an integer
as value (the counter):
- GetNumResources(type) returns the counter value per type.
- GetNumResoourceTypes() returns the size of the map.

Moreover the methods have been exported into class SystemView.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>