[PP_LINUX] Add parsing of CPU QUOTA resources
This patch add support to exploit the CFS Scheduler BANDWIDTH controller, which
is now used to parse the amount of CPU time is available for _each_ CPU
within a node/cluster.

NOTE: The CPU bandwidth is used to assign the SAME quota to each processor
within the same node/cluster.
This is not the intended normal usage of the cfs_quota_us, but it simplifies
a lot the configuration and should be just enough for our purposes, since
we admit a cluster configuration of each NUMA node, where each CPU is used
up to a certain predefined time quota.

Thus, each CPU will receive a % of CPU time defined by:
CPU_QUOTA = cpu.cfs_quota_us * 100 / cpu.cfs_period_us

Signed-off-by: Patrick Bellasi <derkling@gmail.com>