[RTLIB_Monitors] Add goal-less Monitors support
This commit introduces the opportunity of using a monitor and its
features without declaring a goal.

These kind of monitors can be useful for example during Design Space
Exploration (DES). Indeed, in this phase it is important to keep track
of some metrics and take some execution statistics. No goal is defined
since in this phase what is checked is the application behaviors, not
its goals achievements.

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>
Reviewed-by: Andrea Di Gesare <andrea.digesare@gmail.com>