[RTLIB_OPM] Add OP manager support
This introduces an Operating Points Manager, which could be used to
manage all the operating points (OP) of an application and order them
according to different metrics.

Each OP is defined by its own application parameters and a set of
metrics that describe the performance/behavior obtained using that set.
Moreover, it is possible to define a sorting order for metrics allowing
to define a full ordering of operating points, useful for runtime
management purposes.  When such a structure is created, it is possible
to move between different operating points according to priorities and
filters given by the user.

Besides these classes, a general filter on operating points is also
provided.  Each filter consist of the name of a metric, a comparison
function and a value (eg. "Power" LessOrEqual 10). Using a list of
filters such as OP_FilterList, it's possible to filter out from a set of
operating poins the unwanted ones.

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>
Reviewed-by: Andrea Di Gesare <andrea.digesare@gmail.com>