[RTLIB_AS-RTRM] Add Application-Specific RTRM
This commit introduces a new class ApplicationRTRM, which provides a
conveniente glue code between the application Monitors/OPManager and the
System-Wide RTRM.

Indeed, thanks to the knowledge given by a set of registered goals and
from the supporto of the OP Manager, this class is able to take some
decisions, e.g. related to a request for resources increase to be issued
to the BarbequeRTRM.

This is still an early version of the class and it just communicates with
Barbeque when an increase of resources for the application is suggested but
next releases will have  more features, also on the communication with the
application itself.

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>
Acked-by: Andrea Di Gesare <andrea.digesare@gmail.com>