[MCT] Add base class for metrics contributes
This is the base class for implementing contributes of a modular metrics
of a scheduling policy. The class provides a public method Compute(), to
be called by the scheduling policy. This must return the result of the
contribute computation, as an index x {0..1}. The input of the method is
a structure EvalEntity_t, embedding the information related to a single
scheduling possibility, i.e. the tern Application/AWM/Cluster.

More in detail, the index expresses how much is desirable (closer to 1)
or not (closer to 0) to schedule the Application, according to the given
AWM, assigning resource in the scope of the Cluster specified. The
evaluation regards the specific metrics contribute. Example of
contributes are power consumption, congestion avoidance, thermal issues,
reconfiguration overheads and so on.

The implementation of a new contribute requires the defintion of a new
derived class, implementing the private virtual method _Compute().

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>