[MCT_Reconf] Reconfiguration metrics contribute
This contribute returns an index of how much effort would require a
reconfiguration of the Application, in the given AWM, using resources from
a specific Cluster. This contribute is inversely proportional to the amount
of resources required.

Moreover the contribute computation takes into account the weight of a
"migration" into a different cluster, through a configurable proportional
parameter.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>