[MCT_Congestion] Congestion metrics contribute
This contribute starts from the current usage level of each resource
required in the given "scheduling/evaluation entity". On this basis,
three regions of usage are distinguished (by calling suitable methods of
base class MetricsContribute): Sub-Saturation, In-Saturation,
Over-Saturation. The new resource usage request could change the usage
region. According to the region resulting from the resource assignment,
the contribute is computed through a filter function which in turn
determines a linear or exponential penalty, or none.

Once again, the closer to 0 is the contribute index, the lower is the
goodness of the scheduling choice from the congestion prevention point
of view.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>