[Tools_PLP] Add PIL Layout parser
This patch add a new tools folder which will host standalone tools (e.g. script)
which are either required to run the BarbqueRTRM, or which can be useful
to simplify some management of this RTRM.

Tools which are REQUIRED to run BarbequeRTRM, should be defined as
"BarbequeTOOLS" components, while accessory tools could be placed under a
"NBarbequeUTILS" components.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>