[SchedPol_YaMS] YaMS modular scheduling
Add a new scheduling policy plugin which name stands for:
"Yet another Multiobjective Scheduler".

This policy is based on the computing of a metrics resulting by a weighted
sum of an extensible set of contributes (checkout class MetricsContribute).
To add other contributes, please checkout comment string "ADD_MCT".

The weights of the single contributes are configurable through the
BarbequeRTRM configuration file.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>