[Tools_Console] Add a simple BBQ control console script
This is a script to open a BBQ control panel which is (for the time being)
suitable to be used just on Full-HD displays.
This console provides an updated view on:
- syslog
- BBQ log file
- BBQ managed resources
- BBQ managed applications

Some terminals are also opened which are suitable to run commands to:
- Start/Stop the BarbequeRTRM daemon
- Start/Stop applications (e.g. the BbqueTestApp)

A set of scripts are also produced in the <out> folder, which can be on hand
for common management tasks.

NOTE: for the time being, check the HARDCODED configuration files which are
defined at the top of the script.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>