[RTLIB_Monitors] Switch from GAP to NAP penalties
This commit changes the way in which GAPs are calculated.  The previous
way consisted more on a erroneous calculation of a relative error than
to a real GAP (also known as Goal Actual Penalty).  Since this is not
most handy value to use, a decision has been taken: using NAPs.

NAP is an acronym for "Normalised Actual Penalty" and it represents the
effort needed to increase the performance to reach a goal that is
currently not achieved. The function used to calculate it is:

		NAP = |(x-G)/(x+G)|

where 'G' is the goal and 'x' the value observed.

It has good characteristics since it has values between 0 and 1 and it
has the same value if G'=1/G and x = 1/x, which is an interesting
characteristic since it gives the same value both for upper-bounded
goals and lower-bounded goals, e.g timeouts and throughputs
respectively.

Coherently with this decisions, all the references to GAP have been changed
to NAP.

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>
Reviewed-by: Patrick Bellasi <derkling@gmail.com>
Acked-by: Andrea Di Gesare <andrea.digesare@gmail.com>