[RTLIB_Monitors] Bug fix in checkGoal
This commit fixes a small bug related to the checkGoal function.
The input/output parameter "gaps" is taken as it is and it is used to do a
push_back each gap computed by the function.
This commit forces a removal of all the content of the vector before using
it, in order to prevent unpredictability of the output.

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>
Reviewed-by: Andrea Di Gesare <andrea.digesare@gmail.com>