[BuildSystem] Force AS-RTRM after RTLib
This patch ensure that the RTLib target is build before the AS-RTRM related
stuff. This is required since Monitors, OPManager and ApplicationRTRM classes
depends on some RTLib service (e.g. they expect some headers file being
deployed).

The BbqueEXC header is deployed under the building folder, right under the
include/bbque directory where it is expected to be at run-time. This
allows to keep the same sources for the compilation of both AS-RTRM library
and its client applications.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>