[PP_LINUX] Temporary patch to workaround disabled CFS quota management
This is a (perhaps) temporary patch to allow BBQ to run on a 3.2 Linux kernel
which have CPU controller enabled but with CFS Quota management not compiled.

This should be a requirements for a properly usage of BBQ on a >=3.2 kernel,
otherwise a better run-time capabilities identification support should be
put in place.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>