[AWM] Fix total resource checking
When AddResourceUsage() is called by the recipe loader, a check required
regards the amount of resource usage. This obviously should not overcome
the total amount available. This check was previously done by performing
the query on the resource path, instead of the resource template path.
The second indeed allows to perform the query considering the sum of the
total amount of each resource of the same "class".

Therefore the path considered is no more something like
"tile0.cluster1.pe2", but the template version "tile.cluster.pe". This
because the current version of the framework has introduced the possibility
of partition the usage request among the multiple resource of the same
class. For instance, a request of 120% of cpu quota for processing elements
would be satifisfied by gathering 2 processing elements.

Signed-off-by: Giuseppe Massari <joe.massanga@gmail.com>