[RTLIB_Monitors] Bugfix initialization
This commit fixes a bug caused related to the initialization of the Time
and Throughput monitors. The constructor of the classes TimeMonitor and
ThroughputMonitor has been updated in order to avoid to have the boolean
variable "started" unset before using the basic monitors functionalities
(the ones without windows of data and/or goals).

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>
Reviewed-by: Andrea Di Gesare <andrea.digesare@gmail.com>