[BuildSystem] Link bbque_rtlib to bbque_monitors
The RTLib is linked with the Monitors library since they provide now the
ApplicationRTRM class, which is based on RTLib services.

Without this patch the monitors library compiles fine but applications
using it could not be properly linked, if the RTLib is not explicitely
indicated as a terget dependency. This patch should simplify CMakeFiles
of monitors clinent applications.

Signed-off-by: Vincenzo Consales <vincenzo.consales@gmail.com>
Signed-off-by: Patrick Bellasi <derkling@gmail.com>