[PP_LINUX] Bugfix task assignment
To avoid undesired application killing once a task is moved into a CGroup
is better to move the task only after the group has been properly configured.
Thus, due to internals on the libcg, it is safer to use two kenrel updates
calls: one to configure the CGroup and a _second_ one to move the task
into.

Signed-off-by: Patrick Bellasi <derkling@gmail.com>